+++ zone/spell_effects.cpp
@@ -901,7 +901,7 @@ bool Mob::SpellEffect(Mob* caster, uint16 spell_id, float partial)
 						action->source = caster ? caster->GetID() : GetID();
 						action->level = 65;
 						action->instrument_mod = 10;
-						action->sequence = (GetHeading() * 12345 / 2);
+						action->sequence = (uint32)(GetHeading() * 12345 / 2);
 						action->type = 231;
 						action->spell = spell_id;
 						action->buff_unknown = 4;
@@ -950,7 +950,7 @@ bool Mob::SpellEffect(Mob* caster, uint16 spell_id, float partial)
 								action->source = caster ? caster->GetID() : GetID();
 								action->level = 65;
 								action->instrument_mod = 10;
-								action->sequence = (GetHeading() * 12345 / 2);
+								action->sequence = (uint32)(GetHeading() * 12345 / 2);
 								action->type = 231;
 								action->spell = spell_id;
 								action->buff_unknown = 4;
@@ -986,7 +986,7 @@ bool Mob::SpellEffect(Mob* caster, uint16 spell_id, float partial)
 							action->source = caster ? caster->GetID() : GetID();
 							action->level = 65;
 							action->instrument_mod = 10;
-							action->sequence = (GetHeading() * 12345 / 2);
+							action->sequence = (uint32)(GetHeading() * 12345 / 2);
 							action->type = 231;
 							action->spell = spell_id;
 							action->buff_unknown = 4;
